# docs/swagger.yaml
basePath: /
definitions:
...
  response.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  response.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.APIError'
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.Meta:
    properties:
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      timestamp:
        description: Useful for frontend debugging / UX
        type: string
    type: object
  response.PaginationMeta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  workers.CreateRequest:
    properties:
      address:
        type: string
      department:
        maxLength: 100
        type: string
      email:
        maxLength: 255
        type: string
      hire_date:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      notes:
        type: string
      phone:
        maxLength: 50
        type: string
      position:
        maxLength: 100
        type: string
      salary:
        minimum: 0
        type: number
    required:
    - name
    type: object
  workers.ListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/workers.Worker'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  workers.UpdateRequest:
    properties:
      address:
        type: string
      department:
        maxLength: 100
        type: string
      email:
        maxLength: 255
        type: string
      hire_date:
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 255
        minLength: 1
        type: string
      notes:
        type: string
      phone:
        maxLength: 50
        type: string
      position:
        maxLength: 100
        type: string
      salary:
        minimum: 0
        type: number
    type: object
  workers.Worker:
    properties:
      address:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      department:
        type: string
      email:
        type: string
      hire_date:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      notes:
        type: string
      phone:
        type: string
      position:
        type: string
      salary:
        type: number
      updated_at:
        type: string
    type: object
info:
  contact: {}
  description: Market backend (products, stock, income, sales)
  title: Hezzet Market API
  version: "1.0"
paths:
...
  /workers:
    get:
      description: Get paginated list of workers. By default returns only active;
        set active_only=false to include inactive.
      parameters:
      - description: Page (default 1)
        in: query
        name: page
        type: integer
      - description: Limit (default 10, max 200)
        in: query
        name: limit
        type: integer
      - description: Skip (legacy, default 0)
        in: query
        name: skip
        type: integer
      - description: Search by name, phone, or email
        in: query
        name: search
        type: string
      - description: Only active workers (default true)
        in: query
        name: active_only
        type: boolean
      - description: Order by field (name, created_at)
        enum:
        - name
        - created_at
        in: query
        name: order_by
        type: string
      - description: Order direction (asc/desc)
        enum:
        - asc
        - desc
        in: query
        name: order_direction
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/workers.ListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: List workers
      tags:
      - Workers
    post:
      consumes:
      - application/json
      description: Create a new worker (store employee)
      parameters:
      - description: Worker data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/workers.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/workers.Worker'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create worker
      tags:
      - Workers
  /workers/{id}:
    delete:
      description: 'Soft delete worker: sets deleted_at=now and is_active=false'
      parameters:
      - description: Worker ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete worker (soft)
      tags:
      - Workers
    get:
      description: Get a single worker by ID (404 only if deleted)
      parameters:
      - description: Worker ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/workers.Worker'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get worker by ID
      tags:
      - Workers
    patch:
      consumes:
      - application/json
      description: Update worker fields
      parameters:
      - description: Worker ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/workers.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/workers.Worker'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update worker
      tags:
      - Workers
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"